\name{economagicCategories}


\alias{economagicCategories}


\title{Categries of symbols from Economagic}


\description{
    
     Lists categories of symbols from www.economagic.com.
   
}


\usage{ 
economagicCategories()
}


\examples{
## economagicCategories -
   economagicCategories()
}


\keyword{data}

