\name{forecastsCategories}


\alias{forecastsCategories}


\title{Categries of symbols from Forecasts}


\description{
    
     Lists categories of symbols from www.forecasts.org.
   
}


\usage{ 
forecastsCategories()
}


\examples{
## forecastsCategories -
   forecastsCategories()
  
## Category Attributes -  
   attr(forecastsCategories(), "control")
}


\keyword{data}

