\name{show-methods}
\alias{show-methods}

\docType{methods}

%\alias{show,ANY-method}
\alias{show,fWEBDATA-method}


\title{WEBDATA Download Show Methods}

\description{
  Show methods for WEBDATA downloads. 
}

\section{Methods}{
  \describe{

    \item{object = "fWEBDATA"}{
      Print function for objects of class \code{"fWEBDATA"}.
    }   
  }
}

\references{
  Diethelm Wuertz, Yohan Chalabi, and Andrew Ellis (2010);
  \emph{Financial Market Data for R/Rmetrics}, 
  Rmetrics eBook, Rmetrics Association and Finance Online, Zurich,
  www.rmetrics.org.
}

\keyword{data}
