\name{provider-Listings}


\alias{providerListings}

\alias{amexListing}
\alias{h15Listing}
\alias{nasdaqListing}
\alias{nyseListing}
\alias{oandaListing}
\alias{stoxxListing}
\alias{swxListing}


\title{Provider Listing of Symbols and Descriptions} 


\description{

    CSV files with provider listings of symbols, descriptions
    ands related information.
    
    The listings include those from:
    
    \tabular{ll}{
        amexListing      \tab Listing from the American Stock Exchange\cr
        h15Listing       \tab Listing from the FED H15 Report\cr
        nasdaqListing    \tab Listing from the NASDAQ Stock Market\cr
        nyseListing      \tab Listing from the New York Stock Exchange\cr
        oandaListing     \tab Listing of OANDAs Foreign Exchange Rates\cr
        stoxxListing     \tab Listing of STOXX Indices\cr
        swxListing       \tab Listing from the Swiss Stock Exchange\cr
    }
   
}


\format{
    
    All files are given in CSV Excel spreadsheet format. 
    The delimiter is a semicolon. 
    
}


\references{

    Diethelm Wuertz, Yohan Chalabi, and Andrew Ellis, (2010);
    \emph{Financial Market Data for R/Rmetrics}, 
    Rmetrics eBook, Rmetrics Association and Finance Online, Zurich,
    www.rmetrics.org.
    
}
 
 
\keyword{datasets}

