\name{providerListings}


\alias{providerListings}

\alias{amexListing}
\alias{h15Listing}
\alias{nasdaqListing}
\alias{nyseListing}
\alias{oandaListing}
\alias{stoxxListing}
\alias{swxListing}


\title{Provider Listing of symbols and descriptions} 


\description{

    CSV files with provider listings of symbols, descriptions
    ands related information
   
}


\format{
    
    All files are in CSV Excel spreadsheet format. The
    delimiter is a semicolon. 
    
}
 
 
\keyword{datasets}

