\name{yahooExchanges}


\alias{yahooExchanges}


\title{Import Market Data from Yahoo}


\description{
    
     Lists exchanges and symbol modifications supported by
     Yahoo's download service.
   
}


\usage{ 
yahooExchanges()
}


\examples{
## yahooExchanges -
   # yahooExchanges()
}


\keyword{data}

