\name{fImport-package}


\alias{fImport-package}


\alias{fImport}


\docType{package}


\title{Import Data Package}


\description{

    Package of functions to download data from
    the internet or from other sources.
}


\details{

    \tabular{ll}{
        Package:    \tab fImport\cr
        Type:       \tab Package\cr
        Version:    \tab 270.73\cr
        Date:       \tab 2008\cr
        License:    \tab GPL Version 2 or later\cr
        Copyright:  \tab (c) 1999-2008 Diethelm Wuertz and Rmetrics Foundation\cr
        URL:        \tab \url{http://www.rmetrics.org}
    }
    
}
    
  
\section{Overview:}{

    The package makes functions available to download financial market
    data from the internet. Functions are availalbe for the 
    follwing web sites:
    
    1. Data from www.economagic.com\cr
    2. Data from research.stlouisfed.org\cr
    3. Data from chart.yahoo.com
    4. Data from www.oanda.com
    
    Functions for the following functionalities are available:
    
    1. *Series\cr
    2. *Import\cr
    
}    
         
   
\keyword{data} 
       
    