\name{economagicListing}


\alias{economagicListing}


\title{Categries of symbols from Economagic}


\description{
    
     Lists categries of symbols from www.economagic.com
   
}


\arguments{

    \item{category}{
        a character string with filename, usually
        having extension ".csv", where to save the downloaded data.
        }
    \item{abbreviate} {
        ...
        }
}


\examples{
## economagicListing -
}


\keyword{data}

